Create README.md
Boricha TVET College
Property/project Valuatore,Instructore
Duration : 01 October,2019 - still now
As a Property Valuer, My main responsibility will be to correctly identify the true
value of properties and the price they need to be marketed at to generate a successful and fast
sale.And as the instructor teaches in discipline areas in which speci c practical training and
competence. As such, the instructor is expected to plan, organize, and teach in a manner that
encourages Christian development in harmony with the school’s doctrinal position, mission, and
values.